requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
sentence-transformers
faiss-cpu
scikit-learn
nltk